Change strategy from blank state to no relationship types, improved the code and removed comments
