+++ nightmare/helpers/insertFixtures.js
@@ -1,9 +1,9 @@
 import { spawn } from 'child-process-promise';
 
-export default async () => {
+export default async (includeRelationships = true) => {
     try {
-      await spawn('yarn', ['e2e-restore-fixtures'], { capture: ['stdout', 'stderr'] });
+      await spawn('yarn', [`e2e-restore-fixtures${includeRelationships ? '' : '-no-rel'}`], { capture: ['stdout', 'stderr'] });
     } catch (e) {
       process.exit(1);
     }
-}
+};
