+++ package.json
@@ -51,6 +51,7 @@
     "e2e": "jest --projects nightmare/jest.e2e.config.js -i --verbose true",
     "e2e-show": "export SHOW_E2E=1; yarn e2e",
     "e2e-restore-fixtures": "cd nightmare/fixtures;./restore.sh",
+    "e2e-restore-fixtures-no-rel": "cd nightmare/fixtures;./restore_no_rel.sh",
     "eslint": "./node_modules/.bin/eslint app --quiet --ext js,ts,tsx",
     "prettier": "./node_modules/.bin/prettier --loglevel warn --write --require-pragma 'app/**/*.{js,ts,tsx}'"
   },
